Renamed set_up_calculation -> sites_data_importer
