+++ tests/test_sites_data_importer.py
@@ -1,14 +1,14 @@
 import unittest
 from unittest.mock import Mock, patch, mock_open, call
-from pyscses.set_up_calculation import (
-    sites_data_from_file,
-    cluster_similar_sites_data)
 from pyscses.site_data import SiteData, InputFormatError
+from pyscses.sites_data_importer import (sites_data_from_file,
+    cluster_similar_sites_data)
+
 
 class TestSetUpCalculation(unittest.TestCase):
 
-    @patch('pyscses.set_up_calculation.SiteData.from_input_string')
-    @patch('pyscses.set_up_calculation.cluster_similar_sites_data')
+    @patch('pyscses.sites_data_importer.SiteData.from_input_string')
+    @patch('pyscses.sites_data_importer.cluster_similar_sites_data')
     def test_sites_data_from_file_calls_SiteData_from_input_string(self,
         mock_cluster_similar_sites_data,
         mock_from_input_string):
@@ -25,8 +25,8 @@ class TestSetUpCalculation(unittest.TestCase):
                           for line in sites_input_data.split("\n") if line]
         mock_from_input_string.assert_has_calls(expected_calls)
 
-    @patch('pyscses.set_up_calculation.SiteData.from_input_string')
-    @patch('pyscses.set_up_calculation.cluster_similar_sites_data')
+    @patch('pyscses.sites_data_importer.SiteData.from_input_string')
+    @patch('pyscses.sites_data_importer.cluster_similar_sites_data')
     def test_sites_data_from_file_returns_SiteData_sorted_by_x(self,
         mock_cluster_similar_sites_data,
         mock_from_input_string):
@@ -41,9 +41,9 @@ class TestSetUpCalculation(unittest.TestCase):
             sites_data = sites_data_from_file(filename='sites.dat')
         self.assertEqual(sites_data, mock_site_data[::-1])
 
-    @patch('pyscses.set_up_calculation.SiteData.from_input_string')
-    @patch('pyscses.set_up_calculation.SiteData.input_string_is_valid_syntax')
-    @patch('pyscses.set_up_calculation.cluster_similar_sites_data')
+    @patch('pyscses.sites_data_importer.SiteData.from_input_string')
+    @patch('pyscses.sites_data_importer.SiteData.input_string_is_valid_syntax')
+    @patch('pyscses.sites_data_importer.cluster_similar_sites_data')
     def test_sites_data_from_file_validates_input_data(self,
         mock_cluster_similar_sites_data,
         mock_input_string_is_valid_syntax,
@@ -61,9 +61,9 @@ class TestSetUpCalculation(unittest.TestCase):
         expected_calls = [call(line) for line in sites_input_data.split("\n") if line]
         mock_input_string_is_valid_syntax.assert_has_calls(expected_calls)
 
-    @patch('pyscses.set_up_calculation.SiteData.from_input_string')
-    @patch('pyscses.set_up_calculation.SiteData.input_string_is_valid_syntax')
-    @patch('pyscses.set_up_calculation.cluster_similar_sites_data')
+    @patch('pyscses.sites_data_importer.SiteData.from_input_string')
+    @patch('pyscses.sites_data_importer.SiteData.input_string_is_valid_syntax')
+    @patch('pyscses.sites_data_importer.cluster_similar_sites_data')
     def test_sites_data_from_file_raises_InputFormatError(self,
         mock_cluster_similar_sites_data,
         mock_input_string_is_valid_syntax,
@@ -80,9 +80,9 @@ class TestSetUpCalculation(unittest.TestCase):
             with self.assertRaises(InputFormatError):
                 sites_data = sites_data_from_file(filename='sites.dat')
 
-    @patch('pyscses.set_up_calculation.SiteData.from_input_string')
-    @patch('pyscses.set_up_calculation.SiteData.input_string_is_valid_syntax')
-    @patch('pyscses.set_up_calculation.cluster_similar_sites_data')
+    @patch('pyscses.sites_data_importer.SiteData.from_input_string')
+    @patch('pyscses.sites_data_importer.SiteData.input_string_is_valid_syntax')
+    @patch('pyscses.sites_data_importer.cluster_similar_sites_data')
     def test_sites_data_from_file_clusters_similar_sites(self,
         mock_cluster_similar_sites_data,
         mock_input_string_is_valid_syntax,
@@ -100,9 +100,9 @@ class TestSetUpCalculation(unittest.TestCase):
         mock_cluster_similar_sites_data.assert_called_with(sites_data=mock_site_data,
             distance_threshold=1e-10)
 
-    @patch('pyscses.set_up_calculation.SiteData.from_input_string')
-    @patch('pyscses.set_up_calculation.SiteData.input_string_is_valid_syntax')
-    @patch('pyscses.set_up_calculation.cluster_similar_sites_data')
+    @patch('pyscses.sites_data_importer.SiteData.from_input_string')
+    @patch('pyscses.sites_data_importer.SiteData.input_string_is_valid_syntax')
+    @patch('pyscses.sites_data_importer.cluster_similar_sites_data')
     def test_sites_data_from_file_uses_explicit_charges_if_site_charge_is_true(self,
         mock_cluster_similar_sites_data,
         mock_input_string_is_valid_syntax,
@@ -123,9 +123,9 @@ class TestSetUpCalculation(unittest.TestCase):
         self.assertEqual(sites_data[0].valence, -2.0)
         self.assertEqual(sites_data[1].valence, +1.0)
 
-    @patch('pyscses.set_up_calculation.SiteData.from_input_string')
-    @patch('pyscses.set_up_calculation.SiteData.input_string_is_valid_syntax')
-    @patch('pyscses.set_up_calculation.cluster_similar_sites_data')
+    @patch('pyscses.sites_data_importer.SiteData.from_input_string')
+    @patch('pyscses.sites_data_importer.SiteData.input_string_is_valid_syntax')
+    @patch('pyscses.sites_data_importer.cluster_similar_sites_data')
     def test_sites_data_from_file_sets_site_charges_to_zero_if_site_charge_is_false(self,
         mock_cluster_similar_sites_data,
         mock_input_string_is_valid_syntax,
@@ -146,9 +146,9 @@ class TestSetUpCalculation(unittest.TestCase):
         self.assertEqual(sites_data[0].valence, 0.0)
         self.assertEqual(sites_data[1].valence, 0.0)
 
-    @patch('pyscses.set_up_calculation.SiteData.from_input_string')
-    @patch('pyscses.set_up_calculation.SiteData.input_string_is_valid_syntax')
-    @patch('pyscses.set_up_calculation.cluster_similar_sites_data')
+    @patch('pyscses.sites_data_importer.SiteData.from_input_string')
+    @patch('pyscses.sites_data_importer.SiteData.input_string_is_valid_syntax')
+    @patch('pyscses.sites_data_importer.cluster_similar_sites_data')
     def test_sites_data_from_file_sets_site_charges_to_zero_as_default(self,
         mock_cluster_similar_sites_data,
         mock_input_string_is_valid_syntax,
